feat: satellites and solos are now injected as cardinal one subtrees
Satellites and solos were previously injected in clusters before duplications
creation. This method tend to affect intrinsically poorly defined genes in the
wrong places, hampering the growth of duplications and resulting in aberrant,
poorly supported duplications by taking the place of other genes.

These genes are now placed only at the end of the tree reconstruction process.
It does not necessarily increase their quality of positioning, but it favors the
creation of sensible duplicated arms.